update command merged into version
